+++ protected/controllers/OrganizationController.php
@@ -155,6 +155,8 @@ class OrganizationController extends Controller
 
 		$tabs = self::composeOrganizationViewTabs($model, $realm);
 
+		Yii::app()->esLog->log(sprintf("viewed Organization '%s'", $model->title), 'organization', array('trigger' => 'OrganizationController::actionView', 'model' => 'Organization', 'action' => 'view', 'id' => $model->id));
+
 		$this->render('view', array(
 			'model' => $model,
 			'organization' => $model,
@@ -279,7 +281,7 @@ class OrganizationController extends Controller
 
 				Notice::flash(Yii::t('notice', "{title}' updated successfully", ['{title}' => $model->title]), Notice_SUCCESS);
 
-				$log = Yii::app()->esLog->log(sprintf("updated organization '%s'", $model->title), 'organization', array('trigger' => 'OrgranizationController::actionUpdate', 'model' => 'Organization', 'action' => 'update', 'id' => $model->id, 'organizationId' => $model->id));
+				$log = Yii::app()->esLog->log(sprintf("updated Organization '%s'", $model->title), 'organization', array('trigger' => 'OrgranizationController::actionUpdate', 'model' => 'Organization', 'action' => 'update', 'id' => $model->id, 'organizationId' => $model->id));
 
 				if (!empty($returnUrl)) {
 					$this->redirect(urldecode($returnUrl));
@@ -472,7 +474,7 @@ class OrganizationController extends Controller
 			$notifMaker = NotifyMaker::user_hub_revokeEmailAccess($copy);
 			HUB::sendEmail($copy->user_email, $copy->user_email, $notifMaker['title'], $notifMaker['content']);
 
-			$log = Yii::app()->esLog->log(sprintf("deleted '%s' access to organization '%s'", $copy->user_email, $copy->organization->title), 'organization', array('trigger' => 'OrgranizationController::actionDeleteOrganization2Email', 'model' => 'Organization', 'action' => 'deleteOrganization2Email', 'id' => $copy->organization->id, 'organizationId' => $copy->organization->id));
+			$log = Yii::app()->esLog->log(sprintf("deleted '%s' access to Organization '%s'", $copy->user_email, $copy->organization->title), 'organization', array('trigger' => 'OrgranizationController::actionDeleteOrganization2Email', 'model' => 'Organization', 'action' => 'deleteOrganization2Email', 'id' => $copy->organization->id, 'organizationId' => $copy->organization->id));
 
 			if ($realm == 'cpanel') {
 				Notice::flash(Yii::t('notice', "Successfully revoke request from email '{email}' to join '{organizationTitle}'", ['{email}' => $copy->user_email, '{organizationTitle}' => $copy->organization->title]), Notice_SUCCESS);
@@ -538,7 +540,7 @@ class OrganizationController extends Controller
 
 				Notice::flash(Yii::t('notice', "Successfully added email '{email}' to organization '{organizationTitle}'", ['{email}' => $model->user_email, '{organizationTitle}' => $model->organization->title]), Notice_SUCCESS);
 
-				$log = Yii::app()->esLog->log(sprintf("added '%s' access to organization '%s'", $model->user_email, $organization->title), 'organization', array('trigger' => 'OrgranizationController::actionAddOrganization2Email', 'model' => 'Organization', 'action' => 'addOrganization2Email', 'id' => $organization->id, 'organizationId' => $organization->id));
+				$log = Yii::app()->esLog->log(sprintf("added '%s' access to Organization '%s'", $model->user_email, $organization->title), 'organization', array('trigger' => 'OrgranizationController::actionAddOrganization2Email', 'model' => 'Organization', 'action' => 'addOrganization2Email', 'id' => $organization->id, 'organizationId' => $organization->id));
 			} else {
 				Notice::flash(Yii::t('notice', "Failed to add email '{email}' to organization '{organizationTitle}'", ['{email}' => $model->user_email, '{organizationTitle}' => $model->organization->title]), Notice_ERROR);
 			}
@@ -580,6 +582,8 @@ class OrganizationController extends Controller
 			}
 
 			Notice::flash(Yii::t('notice', "Successfully changed status to '{status}'", ['{status}' => $model->formatEnumStatus($model->status)]), Notice_SUCCESS);
+
+			$log = Yii::app()->esLog->log(sprintf("changed '%s' access to '%s' for Organization '%s'", $model->user_email, $model->renderStatus(), $model->organization->title), 'organization', array('trigger' => 'OrgranizationController::actionToggleOrganization2EmailStatus', 'model' => 'Organization', 'action' => 'toggleOrganization2EmailStatus', 'id' => $model->organization->id, 'organizationId' => $model->organization->id));
 		}
 
 		if ($scenario === 'join') {
@@ -742,7 +746,7 @@ class OrganizationController extends Controller
 
 		if ($success) {
 			/* save history of merging */
-			HUB::createOrganizationMergeHistory($source, $target);
+			HubOrganization::createOrganizationMergeHistory($source, $target);
 
 			Notice::page(Yii::t('backend', "Successfully merged '{source}' into '{target}'", ['{source}' => $source->title, '{target}' => $target->title]), Notice_SUCCESS, array('url' => $this->createUrl('organization/view', array('id' => $target->id))));
 		} else {
@@ -783,7 +787,9 @@ class OrganizationController extends Controller
 			//$tmp = HubOrganization::getOrganizationAllActive(1);
 			//$organizations = $tmp['items'];
 			foreach ($organizations as $organization) {
-				$organization->save();
+				if ($organization->save()) {
+					Yii::app()->esLog->log(sprintf("housekeeping for Organization '%s'", $organization->title), 'organization', array('trigger' => 'OrganizationController::actionHousekeeping', 'model' => 'Organization', 'action' => 'housekeeping', 'id' => $organization->id));
+				}
 			}
 		}
 
@@ -885,28 +891,5 @@ class OrganizationController extends Controller
 		}
 
 		return $tabs;
-	}
-
-	/**
-	 * Shows S3 sore of particular startup.
-	 */
-	public function actionScore($id)
-	{
-		// Get name of organization
-		$this->pageTitle = Yii::t('app', 'View Organization');
-		$model = $this->loadModel($id);
-		$this->pageTitle = Yii::t('app', "View '{OrganizationTitle}' Score", array('{OrganizationTitle}' => $model->title));
-		$org_name = $model->title;
-
-		$application = HubAtas::getApplicationFromStartupName($org_name);
-		$application_id = $application['map_application_id'];
-		$this->render('score', array(
-			'model' => $model,
-			'organization' => $model,
-			'application_id' => $application_id
-		));
-		// Get startup_id
-		// Get map_application id
-		// $this->render('score', array());
 	}
 }